Straight Launch to Home works!
I have the smoothest brain imaginable
