+++ RetireSimple.App/.gitignore
@@ -92,3 +92,4 @@ typings/
 out/
 
 resources/
+backend/
