+++ RetireSimple.App/src/index.html
@@ -13,7 +13,7 @@
 		</div>
 	</body>
 
-	<script>
+	<!-- <script>
 		let intervalId = setInterval(() => {
 			//Test if web server is running
 			fetch('http://localhost:5000/api/Heartbeat', {method: 'GET'}).then((response) => {
@@ -28,6 +28,6 @@
 				}
 			});
 		}, 5000);
-	</script>
+	</script> -->
 </html>
