+++ RetireSimple.App/webpack.plugins.ts
@@ -1,6 +1,8 @@
 import type IForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
 // eslint-disable-next-line import/default
 import CopyPlugin from 'copy-webpack-plugin';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const noop = require('noop-webpack-plugin');
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const ForkTsCheckerWebpackPlugin: typeof IForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
@@ -9,15 +11,16 @@ export const plugins = [
 	new ForkTsCheckerWebpackPlugin({
 		logger: 'webpack-infrastructure',
 	}),
-	new CopyPlugin({
-		patterns: [
-			{
-				from: '**/*',
-				to: '[path][name].[ext]',
-				context: 'resources',
-				globOptions: {dot: true, ignore: ['**/node_modules/**']},
-			},
-		], // eslint-disable-line @typescript-eslint/no-unsafe-member-access
-	}),
+	process.env.NODE_ENV === 'production'
+		? new noop()
+		: new CopyPlugin({
+				patterns: [
+					{
+						from: 'backend/**',
+						to: '.',
+						globOptions: {dot: true, ignore: ['**/node_modules/**']},
+					},
+				],
+		}),
 ];
